ft_memset

+++ Libft/Part1-libc/ft_memset.c
@@ -14,10 +14,10 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int i;
+	size_t i;
 
 	i = 0;
-	while (i < (int)len)
+	while (i < len)
 	{
 		b[i] = (unsigned char)c;
 		i++;
